move the index module to index package
