+++ kifi-backend/modules/search/app/com/keepit/dev/SearchDevModule.scala
@@ -9,6 +9,7 @@ import com.keepit.common.store.SearchDevStoreModule
 import com.keepit.inject.CommonDevModule
 import com.keepit.search.spellcheck.SpellCorrectorModule
 import com.keepit.search.tracker.DevTrackingModule
+import com.keepit.search.index.DevIndexModule
 
 case class SearchDevModule() extends SearchModule(
